requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beanie==1.29.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fastapi==0.95.0
fonttools==4.55.4
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httptools==0.6.4
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
lazy-model==0.2.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
motor==3.6.1
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
pydantic==1.10.21
pydantic_core==2.27.2
pymongo==4.9.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.26.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.21.1
uvloop==0.21.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
wheel==0.45.1
yarl==1.18.3
yfinance==0.2.52
